+++ qiskit/quantum_info/synthesis/two_qubit_decompose.py
@@ -40,7 +40,7 @@ from qiskit.circuit.library.standard_gates import CXGate, RXGate, RYGate, RZGate
 from qiskit.exceptions import QiskitError
 from qiskit.quantum_info.operators import Operator
 from qiskit.quantum_info.synthesis.weyl import weyl_coordinates, transform_to_magic_basis
-from qiskit.quantum_info.synthesis.one_qubit_decompose import (
+from qiskit.synthesis.one_qubit.one_qubit_decompose import (
     OneQubitEulerDecomposer,
     DEFAULT_ATOL,
 )
