+++ qiskit/synthesis/__init__.py
@@ -94,6 +94,17 @@ Unitary Synthesis
 
 .. autofunction:: qs_decomposition
 
+One-qubit Synthesis
+===================
+
+.. autosummary::
+   :toctree: ../stubs/
+
+   OneQubitEulerDecomposer
+
+Two-qubits Synthesis
+====================
+
 """
 
 from .evolution import (
@@ -132,3 +143,4 @@ from .stabilizer import synth_stabilizer_layers, synth_stabilizer_depth_lnn
 from .discrete_basis import SolovayKitaevDecomposition, generate_basic_approximations
 from .qft import synth_qft_line
 from .unitary.qsd import qs_decomposition
+from .one_qubit import OneQubitEulerDecomposer
